pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dropspy"
version = "0.1.0"
description = "dropspy"
dependencies = [
    "pytest",
    "pytest-asyncio",
    "colorlog",
    "pyyaml",
    "telethon",
    "python-dotenv",
    "google-cloud-aiplatform[tokenization]",
    "google-genai",
]

[tool.setuptools.packages.find]
where = ["src"]
